pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "packitos"
authors = [
    { name = "Red Hat", email = "user-cont-team@redhat.com" },
]
description = "A set of tools to integrate upstream open source projects into Fedora operating system."
readme = "README.md"
license = "MIT"
license-files = { paths = ["LICENSE"] }
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dynamic = ["version"]
keywords = [
    "git",
    "packaging",
    "fedora",
    "rpm",
    "dist-git",
]
dependencies = [
    "GitPython",
    "PyYAML",
    "cccolutils",
    "click",
    "copr",
    "lazy_object_proxy",
    "marshmallow",
    "marshmallow-enum",
    "munch",
    "ogr",
    "packaging",
    "python-gnupg",
    "requests",
    "requests-kerberos",
    "specfile",
    "tabulate",
    "bodhi-client >= 7.0.0",
    "koji",
    "rpkg",
    "cachetools",
    "python-fedora",
]

[project.urls]
Homepage = "https://github.com/packit/packit"

[project.optional-dependencies]
testing = [
    "pytest",
    "pytest-cov",
    "pytest-timeout",
    "flexmock",
    "deepdiff",
    "distro",
    "setuptools", # Required for test_upstream.py/test_get_version_macro
    "setuptools-scm", # Required for tests using python-ogr.spec
    "setuptools-scm-git-archive", # Required for tests using python-ogr.spec
]
dev = [
    "packitos[testing]",
    "pre-commit",
]

[project.scripts]
packit = "packit.cli.packit_base:packit_base"

[tool.hatch.version]
source = "vcs"
raw-options.version_scheme = "post-release"

[tool.hatch.build.targets.wheel]
packages = ["packit"]

[tool.pytest.ini_options]
filterwarnings = "ignore::DeprecationWarning"
addopts = '-m "not slow"'
markers = [
    "slow: Slow tests",
]
testpaths = [
    "tests",
]

[tool.ruff]
select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "COM", # flake8-commas
    "E",   # pycodestyle Error
    "F",   # Pyflakes
    "I",   # isort
    "PERF",# Perflint
    "PIE", # flake8-pie
#    "PT",  # flake8-pytest-style, would be nice to enable
    "RET", # flake8-return
    "RUF", # Ruff-specific rules
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle Warning
]
ignore = [
    "B017", # assert-raises-exception
    "B022", # useless-contextlib-suppress
    "RUF003"# Comment contains ambiguous character
]

line-length = 100
target-version = "py39"
